requirements.txt
transformers
python-telegram-bot
torch
diffusers
